RavenDB-21993 - fix so changes api connection doesn't get disposed on "no reply for operation", only a database not exists error. added test. moved tests.
