+++ src/Raven.Client/Http/RequestExecutor.cs
@@ -2463,6 +2463,8 @@ namespace Raven.Client.Http
             internal Action DelayRequest;
 
             internal Action<GetDatabaseTopologyCommand> SetCommandTimeout;
+
+            public ManualResetEventSlim WaitBeforeFetchOperationStatus;
         }
     }
 }
